pyproject.toml
[project]
name = "agent-bootcamp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=1.2.0",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.6",
    "python-dotenv>=1.2.1",
    "tavily-python>=0.7.17",
]
